requirements.txt
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cymem==2.0.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
langcodes==3.2.1
MarkupSafe==2.0.1
murmurhash==1.0.6
numpy==1.21.4
packaging==21.2
pathy==0.6.1
Pillow==8.4.0
preshed==3.0.6
pydantic==1.8.2
pyparsing==2.4.7
requests==2.26.0
smart-open==5.2.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==2.0.2
